Move all game logic to the server side
