+++ src/server/scripts/game.js
