+++ src/server/scripts/hand.js
